requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data processing and analysis
pandas==2.1.1
numpy==1.25.2

# Scheduling and background tasks
schedule==1.2.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Production server
gunicorn==21.2.0

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3